+++ clean.ps1
@@ -1,9 +1,10 @@
-rm -rf vendors/build
-rm -rf vendors/openblas
-rm -f vendors/openblas.zip
-rm -rf vendors/swig
-rm -f vendors/swig.zip
-
 rm -rf faiss/build
-
 git restore faiss
+
+if ($args[0] -eq "all") {
+    rm -rf vendors/build
+    rm -rf vendors/openblas
+    rm -f vendors/openblas.zip
+    rm -rf vendors/swig
+    rm -f vendors/swig.zip
+}
